Add scrolling functionality when clicking tabs. This is VERY hacky and not modular at all, but it (kind of) works (for now anyways).
Lots of weird stuff popping up that needs to be addressed: the contact info sits above the header, regardless of z-index (ugh), and after it overlaps the last two tabs, they lose the ability to be clicked (until one of the others is clicked, then the functionality returns).

When I adjust the content of the elements in Content, I'll need to manually adjust the placement of the scrollIntoView tags, which sucks.

Still no scrolling-to-change-which-tab-is-highlighted either, and I have really no idea how I'm going to do that.